Create folder components/ClaimableBalances and moved claimable balance file into it
