+++ src/components/ClaimableBalances/SuccessfulClaimTransaction.tsx
