+++ src/pages/Dashboard.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { BalanceInfo } from "components/BalanceInfo";
-import { ClaimableBalances } from "components/ClaimableBalances";
+import { ClaimableBalances } from "components/ClaimableBalances/ClaimableBalances";
 import { TransactionHistory } from "components/TransactionHistory";
 import { LiquidityPoolTransactions } from "components/LiquidityPoolTransactions";
 import { logEvent } from "helpers/tracking";
